Update dockerfile, env loading. Add metrics utilities
